refactor(phase-3): flip CamoufoxLauncher impl Common → Pipeline (C8)
WHY
---
Phase 3 (Common ↔ Pipeline unification) of the master decoupling
plan mandates that Pipeline owns every canonical implementation
and Common is reduced to thin back-compat shims. CamoufoxLauncher
was the inverted case — the 117-line impl lived in Common while a
14-line shim sat in Pipeline. This commit flips that relationship
so Pipeline holds the single source of truth and Common keeps a
≤25-line re-export shim for the existing import sites:

  - src/Scrapers/Base/BaseScraperWithBrowser.ts  (prod consumer)
  - src/Tests/E2eMocked/{Discount,Max,VisaCal}.e2e-mocked.test.ts
  - src/Tests/E2eMocked/Helpers/BrowserFixture.ts
  - src/Tests/E2eReal/Tools/CaptureInvalidLogin.ts
  - src/Tests/Unit/Common/CamoufoxLauncherKnobs.test.ts  (drift canary)

After this flip, Probe 3.2 (Pipeline → Common edges, excluding
the allowlisted BrowserConfig) reaches its DoD target of 0.

WHAT
----
* MOVED the 117-line CamoufoxLauncher implementation from
  src/Common/CamoufoxLauncher.ts INTO
  src/Scrapers/Pipeline/Mediator/Browser/CamoufoxLauncher.ts
  (which was a 14-line back-compat shim before this commit).

* CONVERTED src/Common/CamoufoxLauncher.ts into a 25-line shim
  that re-exports the four public symbols (ISRAEL_LOCALE,
  launchCamoufox, buildLaunchOptions, envFlag) plus the EnvFlag
  type from the canonical Pipeline file. No import site needs an
  edit; all eight downstream files keep working unchanged.

* INTRODUCED the EnvFlag brand (Brand<boolean, 'EnvFlag'>) inside
  the canonical Pipeline file so envFlag satisfies Pipeline
  Rule #15 (no primitive returns from exported helpers) while
  remaining assignable to plain-boolean Camoufox launch-option
  fields (humanize / disable_coop / block_webrtc) via the
  brand's contravariant subtype.

* PRESERVED every anti-detect knob and pinned fingerprint
  bit-for-bit — humanize / disable_coop / block_webrtc defaults,
  CAMOUFOX_* env-flag bisect surface, Windows os pin, locale =
  he-IL, window/screen = 1920x1080. The CamoufoxLauncherKnobs
  drift canary continues to pass unchanged against the relocated
  impl, proving zero behavioural drift across the FLIP.

* RELATIVE-PATH adjustment: the canonical Pipeline file imports
  DESKTOP_VIEWPORT_{WIDTH,HEIGHT} + ISRAEL_LOCALE from
  ../../../../Common/Config/BrowserConfig.js (the BrowserConfig
  file is on the Phase-3 allowlist and stays in Common).

* DOCS — extended two docs pages so the docs-coverage gate's
  cumulative diff (Phase-3 branch vs main) finds whole-word
  matches for every NEW Pipeline export introduced since main:
    - docs/architecture/waf-interceptor.md — new
      "Camoufox launcher knob surface" section names
      launchCamoufox / buildLaunchOptions / envFlag / EnvFlag
      and the CAMOUFOX_* bisect env vars.
    - docs/phases/otp-trigger.md — new "Canonical helpers"
      section names detectOtpScreen / extractPhoneHint /
      findOtpSubmitSelector / clickOtpTriggerIfPresent /
      clickFromCandidates plus OTP_INPUT_CANDIDATES /
      OTP_SUBMIT_CANDIDATES / OTP_TEXT_PATTERNS /
      SMS_TRIGGER_CANDIDATES / PHONE_PATTERN from the canonical
      Pipeline OtpDetector + OtpDetectorConfig modules (these
      were exported in C6 but never picked up by the gate's
      symbol-set diff until C8's run surfaced them).

VALIDATION
----------
* tsc --noEmit : green (no diagnostics)
* eslint src/Common/CamoufoxLauncher.ts
         src/Scrapers/Pipeline/Mediator/Browser/CamoufoxLauncher.ts
  : green (0 errors after simple-import-sort auto-fix on the shim)
* npm run lint:architecture : green (all 15+ architecture rules pass)
* npm run test:pipeline (full Pipeline jest run, 422 suites / 4942
  tests): GREEN. Coverage:
    Statements 97.12% (threshold 97) ✓
    Branches   95.06% (threshold 95) ✓
    Functions  97.24% (threshold 97) ✓
    Lines      98.35% (threshold 98) ✓
  Per-file coverage of the relocated CamoufoxLauncher.ts:
    Statements 100% (11/11)
    Branches   100% (2/2)
    Functions  100% (3/3)
    Lines      100% (10/10)
  Two test suites exercise the moved file:
    - Tests/Unit/Common/CamoufoxLauncherKnobs.test.ts (29 tests)
      drives buildLaunchOptions + envFlag through all
      anti-detect knob defaults + the 7 truthy + 6 falsy env
      string variants per knob.
    - Tests/Unit/Pipeline/Mediator/Browser/CamoufoxLauncher.test.ts
      (3 tests) confirms ISRAEL_LOCALE re-export and the
      launchCamoufox async-function shape.

PROBE DELTAS
------------
* Probe 3.2 (Pipeline → Common edges, allowlist-excluded):
    pre-C8:  1 (Pipeline shim → Common canonical)
    post-C8: 0  ✓ — hits the DoD target locked-in for C5-C9 hold
* Probe 3.1 (Common file count):
    pre-C8:  N  (Common/CamoufoxLauncher = 117L canonical)
    post-C8: N  (Common/CamoufoxLauncher = 25L shim — file kept
                 for back-compat; collapses in a later cleanup
                 once all 8 importers are flipped to Pipeline)
* No change to Probe 3.4 (planned C11 enforcement target).

NEXT IN LADDER
--------------
C9 — Debug UNIFY (~20 Common-side callsites; add overload that
accepts both getDebug(metaUrl) and getDebug(name); allows
src/Common/Debug.ts to become a shim and unlocks
src/Common/Config/DebugConfig.ts deletion).
