+++ docs/architecture/waf-interceptor.md
@@ -151,3 +151,7 @@ No PII flows through these events — only the `kind` enum (`hcaptcha-checkbox` 
 | [`Core/Builder/PipelineBuilderHelpers.ts`](https://github.com/sergienko4/israeli-bank-scrapers/blob/{{BRANCH}}/src/Scrapers/Pipeline/Core/Builder/PipelineBuilderHelpers.ts) | Wires interceptor FIRST in `buildBrowserBaseInterceptors()` |
 | [`Interceptors/WafChallenge/WafChallengeDetector.ts`](https://github.com/sergienko4/israeli-bank-scrapers/blob/{{BRANCH}}/src/Scrapers/Pipeline/Interceptors/WafChallenge/WafChallengeDetector.ts) | Frame URL classifier |
 | [`Interceptors/WafChallenge/HCaptchaCheckboxSolver.ts`](https://github.com/sergienko4/israeli-bank-scrapers/blob/{{BRANCH}}/src/Scrapers/Pipeline/Interceptors/WafChallenge/HCaptchaCheckboxSolver.ts) | Solve recipe |
+
+## Camoufox launcher knob surface
+
+The browser that the interceptor runs inside is built by `launchCamoufox` in `src/Scrapers/Pipeline/Mediator/Browser/CamoufoxLauncher.ts`. The launcher centralises the three Cloudflare-managed-challenge auto-pass knobs (`humanize`, `disable_coop`, `block_webrtc`) and pins a deterministic Windows 1920x1080 fingerprint so banks cannot serve mobile content via screen-size heuristics. CI bisects can flip individual knobs via the `CAMOUFOX_HUMANIZE` / `CAMOUFOX_DISABLE_COOP` / `CAMOUFOX_BLOCK_WEBRTC` env vars without a code edit; the `envFlag` helper parses those env vars into the nominal `EnvFlag` brand. The `buildLaunchOptions` helper bundles every option in one place so a future tuning pass cannot miss a copy. The `Tests/Unit/Common/CamoufoxLauncherKnobs.test.ts` drift canary locks every knob bit-for-bit.
